fix: code format (#181)
* fix: code format

* fix: lint
